+++ CHANGELOG.md
@@ -13,6 +13,16 @@ and this project adheres to **[Semantic Versioning](https://semver.org/spec/v2.0
 
 ---
 
+## [0.1.0] - 2026-08-20
+
+### Added
+
+- Added `save_chart()` for saving Matplotlib `Axes` to a specified path.
+- Added documentation and examples for the create / save / show workflow.
+- Added additional public API docstrings.
+
+---
+
 ## [0.0.2] - 2026-08-16
 
 ## Added
@@ -131,7 +141,8 @@ git push origin :refs/tags/vX.Z.Y
 
 ## Links
 
-[Unreleased]: https://github.com/analytics-toolworks/ml-vizkit/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/analytics-toolworks/ml-vizkit/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/analytics-toolworks/ml-vizkit/releases/tag/v0.1.0
 [0.0.2]: https://github.com/analytics-toolworks/ml-vizkit/releases/tag/v0.0.2
 [0.0.1]: https://github.com/analytics-toolworks/ml-vizkit/releases/tag/v0.0.1